+++ .github/CHANGELOG.md
@@ -2,6 +2,45 @@
 
 <h3>New features since last release</h3>
 
+- The number of shots can now be specified on a temporary basis when evaluating a QNode.
+  [(#1075)](https://github.com/PennyLaneAI/pennylane/pull/1075)
+
+  ```python
+  dev = qml.device('default.qubit', wires=1, shots=10) # default is 10
+  
+  @qml.qnode(dev)
+  def circuit(a):
+      qml.RX(a, wires=0)
+      return qml.sample(qml.PauliZ(wires=0))
+  ```
+  
+  For this, the qnode is called with an additional `shots` keyword argument:
+  
+  ```pycon
+  >>> circuit(0.8)  
+  [ 1  1  1 -1 -1  1  1  1  1  1]
+  >>> circuit(0.8, shots=3)
+  [ 1  1  1] 
+  >>> circuit(0.8)  
+  [ 1  1  1 -1 -1  1  1  1  1  1]
+  ```
+
+- The JAX interface now supports all devices.
+  [(#1076)](https://github.com/PennyLaneAI/pennylane/pull/1076)
+
+   Here is an example of how to use JAX with Cirq:
+
+  ```python
+  dev = qml.device('cirq.simulator', wires=1)
+  @qml.qnode(dev, interface="jax")
+  def circuit(x):
+      qml.RX(x[1], wires=0)
+      qml.Rot(x[0], x[1], x[2], wires=0)
+      return qml.expval(qml.PauliZ(0))
+  weights = jnp.array([0.2, 0.5, 0.1])
+  print(circuit(weights)) # DeviceArray(...)
+  ```
+
 - Added the `ControlledPhaseShift` gate as well as the `QFT` operation for applying quantum Fourier
   transforms.
   [(#1064)](https://github.com/PennyLaneAI/pennylane/pull/1064)
@@ -46,7 +85,15 @@
 
 <h3>Breaking changes</h3>
 
+* If creating a QNode from a quantum function with an argument named `shots`,
+  a `DeprecationWarning` is raised, warning the user that this is a reserved 
+  argument to change the number of shots on a per-call basis.
+
 <h3>Bug fixes</h3>
+
+* Fixes a bug where inverse operations could not be differentiated
+  using backpropagation on `default.qubit`.
+  [(#1072)](https://github.com/PennyLaneAI/pennylane/pull/1072)
 
 <h3>Documentation</h3>
 
@@ -54,7 +101,7 @@
 
 This release contains contributions from (in alphabetical order):
 
-Thomas Bromley, Josh Izaac, Daniel Polatajko
+Thomas Bromley, Josh Izaac, Daniel Polatajko, Chase Roberts, Maria Schuld
 
 # Release 0.14.0 (current release)